requirements.txt
Django>=5.1.7
Pillow==10.0.0
python-dotenv>=1.0.0
django-crispy-forms>=2.1
crispy-bootstrap5>=0.7 